Fix coments from pull request
