--- bin/aeproject/cli-commands/test/test-config.js
@@ -1,70 +1,0 @@
-/*
- * ISC License (ISC)
- * Copyright (c) 2018 aeternity developers
- *
- *  Permission to use, copy, modify, and/or distribute this software for any
- *  purpose with or without fee is hereby granted, provided that the above
- *  copyright notice and this permission notice appear in all copies.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
- *  REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
- *  AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
- *  INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
- *  LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
- *  OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
- *  PERFORMANCE OF THIS SOFTWARE.
- */
-
-const test = require('./test');
-const dir = require('node-dir');
-let path = require("path");
-
-
-const run = async (path = './test', skipCompilation) => {
-  var workingDirectory = getWorkDirectoryPath();
-  var testDirectory = '';
-
-  if (path.includes('.js')) {
-    await test.run([path], skipCompilation);
-
-    return;
-  }
-
-  testDirectory = path;
-
-  if (!path.includes(workingDirectory)) {
-    testDirectory = `${process.cwd()}/${path}`;
-  }
-
-  const files = await getFiles(testDirectory);
-
-  await test.run(files);
-}
-
-const getFiles = async function (testDirectory) {
-  let testsRegex = /.*\.(js|es|es6|jsx|sol)$/
-
-  return new Promise((resolve, reject) => {
-    dir.files(testDirectory, (error, files) => {
-      if (error) {
-        reject(new Error(error));
-
-        return;
-      }
-
-      files = files.filter(function (file) {
-        return file.match(testsRegex) != null;
-      });
-
-      resolve(files);
-    });
-  });
-}
-
-function getWorkDirectoryPath() {
-  return path.join(process.cwd(), "test");
-}
-
-module.exports = {
-  run
-}
